Updated Debugging.
Currently the Printer sends an Home0 Command and waits for ok and an new cmd. After sending an cmd it sends it to the MFU and waits for an ok. Then it jumps to Received OK after waiting for toolchange.

Maybe split up the code into Unload(mm), Load(mm) and Change Tool(index). Then all the control is on the Printer side. This requires more states and maybe some more printerData-Storage (Position, Temp <= Before Change)